rename `generate` function to `generate_per_pool`
