+++ jwst/associations/tests/test_level3_spectrographic.py
@@ -11,7 +11,7 @@ from jwst.associations.tests.helpers import (
 )
 
 from jwst.associations import generate
-from jwst.associations.main import constrain_on_candidates
+from jwst.associations.lib.utilities import constrain_on_candidates
 
 
 class TestLevel3Spec(BasePoolRule):
